Add Exoscale host provider
Exoscale is a Swiss cloud provider and one of the provider currently
supported in Docker Machine. Its addition in Rancher UI follows a logic
similar to Amazon EC2 with a multi-step wizard:

 1. Ask for credentials.
 2. Grab the list of security groups.
 3. Ask which security group to use and propose to create the default
    one if it doesn't exist.
 4. Create the default security group if requested.
 5. Grab the instance profiles and available disk sizes.
 6. Let the user choose a name, instance profile and disk size.

Currently, only one availability zone is available. Therefore, there is
no drop-down list to choose one.

While there is several images available, we enforce the use of Ubuntu
Trusty. It would be possible to let the user choose, but this needs some
work to know which images would work and which wouldn't.

The default ingress rules also include the Docker daemon port. Docker
Machine needs it for the first time installation.