Fix the select render issue in Firefox
if there is a matching value in options,  set the value manually to avoid the Firefox issue https://github.com/emberjs/ember.js/issues/15484

In chrome and firefox, the first option will be selected automatically when the value is not in the content array. But we need to trigger the change event, otherwise the value in the model will not be updated.

https://github.com/rancher/rancher/issues/10242
https://github.com/rancher/rancher/issues/10241