Update Pull Request Template
We've had some discussion recently and checklists seem to be extraneous so they have been removed. I also moved the guidence comments into HTML comments. The original intent was that people would replace that content with the info that was being requested.  That was not happening and it was making PR's a bit tough to read.