Add more context to the title of deletion modals
Adds the resource types "namespace", "cluster", "project" as
additional context to the titles of the respective deletion modals
in order to reduce the likelyhood of a user accidentally deleting
an unintented resource.

We changed the title from "Are you sure you want to delete" to
"Are you sure you want to delete the namespace:" with namespace
being bold. The same is true for cluster and project as well.

rancher/rancher#17619