Add more context to the namespace deletion modal
Adds the resource type "namespace" as additional context to the
title of the namespace deletion modal in order to reduce the
likelyhood of a user accidentally deleting an unintented resource.

We changed the title from "Are you sure you want to delete" to
"Are you sure you want to delete the namespace:" with namespace
being bold.

Resolves #17619