Improve the mobile sub-navigation UX
In the navigation menu when a user attempted to open a sub-menu the
entire navigation menu instantly closed. To resolve this we now listen
to the router willTransition event to close the navigation rather than
a willRender invocation.

The sub-menu was also overlapping other menu items because it had an
absolute position. To resolve this we now give the sub-menu a static
position.

rancher/rancher#21914