Display security options when editing GKE cluster
We want the user to be able to see the security options that were
selected even if they can't be edited when editing the cluster.

We had to extract and infer the selected options given the oauthScopes.
It would be better if our API could more closely reflect our fields.

rancher/rancher#19070