Include Template/Revision in cluster table
We want to make it easier for someone that is managing multiple
clusters to see what template and template revision a cluster
is using. We also want to make it easy for a user to notice
that a cluster isn't using the latest revision.

To achieve this we added the template and revision to a column
in the clusters table to display both. We also indicate add
a warning badge beside a revision when the revision isn't
the latest.

rancher/rancher#22047