Making the cluster template revisionId saveable.
When editing a cluster that was created with cluster template
the cluster template revision couldn't be saved.

The revisionId was stored as a component member variable instead
of as a part of the model. It needed to be stored as part of the
model in order for the NewOrEdit to see the changes and save
them. I went ahead and referenced the model directly everywhere
in the component and removed the component member variable.

rancher/rancher#22920