Make groups visible when editing cluster template
While editing a cluster template when the template was shared with a
group the group was not visible under the Share Template section. We
needed to ensure that the group was visible.

The reason that this was occuring was because we weren't fetching the
principal for the group because we were only fetching when the
userPrincipalId was present and not when the groupPrincipalId was
present.

rancher/rancher#22971