Improve and fix supported storage filtering
Editing was failing when an existing unsupported storage type had
already been used and the ferature flag unsupported-storage-drivers
was false. Adding a volume as part of a workload was also broken
when attempting to use an unsupported storage type.

To remedy this I decided to refactor the code and make the filtering
happen explicitly only for the dropdowns.

I also decided to make feature checking reusable and either to use
by adding a service to check if a flag is enabled.

rancher/rancher#22895
rancher/rancher#22998