check if cluster local auth endpoint from api is string or bool
Tweaking a change made in rancher/ui#3376 to account for both cases. the fix
added would allow the case when deploying from a template but would break for
nontemplate launches when the value would be a bool again. Added a type check to
handle both cases.

The real fix should be changing the types in the `listquestions` api
where the data is being pulled from. The root cause exists there where the type
`boolean` is saved as a string.

rancher/rancher#22972