Use the correct variable to check if localClusterAuthEndpoint avail
I saw this while fixing another issue, if initialized to false we set the
`enableLocalClusterAuthEndpoint` variable as such but that did not propigate
back up the model when in the init lifecycle. Rather than check the model we
should check the local param so we don't run in to these out of sync issues.