Upgrade Ember to 3.12
Remove ember object and use regular object instead

Found durning upgrade to 3.12, I'm not sure why we'd even need to declare an
ember object for this return anyway.

Get the current cluster's id from scope servervice

Found durning upgrade to 3.12, not sure how this ever worked as I looked on live
and it was failing but the system didn't explode on 3.8. In 3.12
transistion.params is undefined and the whole system explodes. We were never
passing the queryparams so they wouldn't have come through. I changed this so we
fetched the current cluster id from the scope service and everything seems to
work correctly.

Ensure that accordion items only try to expand once during the init

Found durning upgrade to 3.12, expanded was being double rendered.

lint fix

test fixes

lint