Remove nav tabs and simplify Istio ux
I commented out the destinationRules and virtualServices tabs from the
istio-project template, changed some verbiage in the remaining nav tabs, and
added to import yaml link to both tabs to make the ux a little less confusing.
In the next release this page will undergo a redesign so I did not remove the
code.

rancher/rancher#23108