dedup import yaml button when istio feature flag is enabled
When adding the vitual services and destination rules are enabeld we were
showing the import yaml button twice. It needs to be visible on both tabs and we
should be able to pass the or to the rbac check so I removed the duplicates and
moved the button to before the potential other buttons in the istio nav.

rancher/rancher#23155