Ensure project-member-row uses same identicon as principal model
project-member-row was using `avatarSrc` on its component which was slightly
different then the same code used in the principal model but really should have
been the same. I got rid of the duplicate code in the component since the model
is the data owner. It really should manage the avatar source building.

I also updated single quotes to doubles in template and refactored the owner
logic since it wasn't doing anything extra that the noUpdate flag was doing.

rancher/rancher#21440