Fix routing issues post ember upgrade
It seems the previous way that the cluster id and project id was fetched from
various addons (alerts and istio) does not work in ember 3.12. After some
investigation it wasn't the corret way to grab those parameters anyway. I've
changed the routes to use the scope service which has the current cluster and
project (if in one) rather than fetch the query params off a parent route
outside the engine.