Restrict Istio member access to groups only
Due to some oddities in the way that Istio currently allows access to members to
the kiali and jaeger ui's we've decided to restrict the member search to groups
only. I've changed the verbiage on the lables to reflect this. Additionally I've
also exposed the searchOnlyGroups param on the principal search component. I did
not expose the user only param because this is not something we're normally
going to be doing.

rancher/rancher#23215