route name is getting modified twice
there is a weird issue in ember engines that resets the route path which can
cause a double render error to be thrown by ember. Easiest fix is to unbind the
param and since this is only dynamic because we're building the list of links
from an array there is no reason for it to be bound.