add default else to mappedVersions itteration when on current version
When we were looping over the versions to be displayed on init there was a
missing else case that would cause a null value to be pushed into the
mappedVersions array and would cause a JS error when setting the current value
at the end of this statement. This could cause the bugs where the page would
"freeze" when swapping back and forth between revisions on cluster creation.

rancher/rancher#23359