reset clusters CTRID when revision changes on edit cluster
At one point we changed the cluster-driver/rke to use the models CTRID instead
of the watched query params. This was fixed in rancher/ui#3379 and while that
sovled the problem of the ticket it didn't change the id on the cluster model so
the changes were not persisted.

rancher/rancher#23383