[2.3.1] RKE Template Bugs (#3445)
* set form-versions value correctly when applying RKE template

We weren't accounting for applying a cluster template when setting the value of
a form version when using a .x based value. In this case we'd set the value of
the kubernetes version to the content of the label which caused all sorts of
issues.

rancher/rancher#23360
rancher/rancher#23361

* add default else to mappedVersions itteration when on current version

When we were looping over the versions to be displayed on init there was a
missing else case that would cause a null value to be pushed into the
mappedVersions array and would cause a JS error when setting the current value
at the end of this statement. This could cause the bugs where the page would
"freeze" when swapping back and forth between revisions on cluster creation.

rancher/rancher#23359

* reset clusters CTRID when revision changes on edit cluster

At one point we changed the cluster-driver/rke to use the models CTRID instead
of the watched query params. This was fixed in rancher/ui#3379 and while that
sovled the problem of the ticket it didn't change the id on the cluster model so
the changes were not persisted.

rancher/rancher#23383

* fix editable on network prov for cluster templates

if you cloned a template with a network provider override available you couldn't
edit the provider as the cluster template creator

* remove arbitrary patch value for windows k8s support

k8s started supporting windows in 1.14, rancher started supporting it in 1.15, I
am not quite sure why the `.3` patch value was being used but we're safe to set
the min version to 1.15.0 as the label says

rancher/rancher#23395