check if k8s version is user override or patchVersion override
rancher/rancher#23413
rancher/rancher#23415

In the event that the cluster template creator selected a k8s version that was a
patch release version when we came in to clone the revision the UI would show
the question as an override in which case the `satisfies` param was dropped. Now
when a clone happens we check to see if its the patch and has satisifies and in
this case we "remove" the override because its not technically one. The logic on
the cluster template save will ensure the new "override" is added on save with
the correct version.