disp k8s ver correctly when editing cluster or clone rev
the check-override-allowed component did not know how to deal with the k8s
version question because of its tri state and how we deal with the patch version
that is an override but not really an override. I added a check to verify the
mode is view and we have the param then display param so we don't initialize the
form-version component which has logic to inject the current version into its
versions dropdown but only if we're new, editing, or cloning.

rancher/rancher#23478
rancher/rancher#23465