Improve kube version inspection on cluster edit
While editing a cluster properly support .x kube version comparisons when
filtering out cluster template revisions.

Coercing a .x version converts it to a .0 which made the revision look like
it was a kube downgrade. By making use of .satisfies when the revision
kube version ends with a '.x' we're now better able to check if the
kube version is a downgrade and filter appropriately.

rancher/rancher#23489