Ensure istio member type is correct on add
When adding a github group to be an instio member we wanted the
appropriate Member Type to be displayed. We wanted 'Organization'
instead of 'group'.

We noticed that addAuthorizedPrincipal was being invoked with a newly
created member object instead of just passing the principal. Since
everything we could inspect code wise expected a principal instead
of a member I went ahead and just passed the principal.

rancher/rancher#23264