Fix EKS cluster creation transition
After pressing the 'create' button of the EKS driver the user was being
transitioned back to step 3 (Vpc & Subnet) rather than waiting on the
final page until the save is complete and returning to the cluster page.

An observer was being triggered by the save process which subsequently
set the step back to 3. To resolve this we will only enter the branch if
there are initialized values that need to be set back to default.

rancher/rancher#23493