refactor global permissions to simplify and not overwrite roles
When global roles were added the front end had to make some assumptions for
non-admin users to ensure the user had login access. This role used to be hidden
but was made visible by default. With this exposed we now check before save if
the user has some kind of login access and throw an error if not. We also no
longer need to hide the custom roles which allows new user default roles to be
displayed without all the crazy logic to decide which mode to show. The old
login-access read only param was removed; This used to map to user-base and we
only displayed it because the role was hidden.

rancher/rancher#23770