start loading the intl lang sooner to ensure loaded b4 first render
because we did not delay the readiness of the application when loading
translations when a user hit `update-setting` route we'd render the page before
the language had finished loading. Moving the language to the before model and
in the finally of the detect phase ensures we have our language loaded before we
remove the loading state and render the main content.

rancher/rancher#23643