Prevent App editing from modifying state on cancel
When editing both single and multi-cluster apps if you modified the
Template Version or the Target Projects and then cancelled it the
underlying store was still modified which then reflected those changes
on other pages like the single and multi-Cluster Apps pages.

To change this I cloned and nested the fields into a nested object named
'editable' and updated the primarySource on willSave for the relevant
targets subfield.

rancher/rancher#21228