Fix sorting on cloud credentials page
While I was cleaning this up I found that groupByRef in cc this case did nothing
so I removed it. This exposed a bug in the sortable-table code where groupByRef
was coming back null but when we tried to get it embers get would fail cause we
passed a null value instead of an empty string. Additionally I found that when
using presorts we were continously pushing into the array on the component, which
never gets reset, so I decided to clone the array.

Additionally we never added the descending key to our presorts so I've added
that.

rancher/rancher#23781