Fix sorting on cloud credentials page
While I was cleaning this up I found that groupByRef in cc this case did nothing
so I removed it. This exposed a bug in the sortable-table code where groupByRef
was coming back null but when we tried to get it, using embers getter, it would
fail because we passed a null value instead of an empty string.

Additionally I found that when using presorts we were continuously pushing into
the sort array on the component, which never gets reset, so I decided to clone
the array. We rarely use preSorts so it was never caught.

Additionally we never add/remove the descending key to our presorts so I've
added that. Without these when we pass the sort to embers sort we'd never
correctly sort any presorts based the descending flag.

rancher/rancher#23828