Notify user of value missing from options
When using an older vsphere node template in the latest version it was
possible that a user entered value would not be a valid option. When
the value isn't a valid option the dropdown has nothing selected. To
improve the UX we now notify the user that the current value is no
longer valid and sask them to select a new value.

rancher/rancher#23920