Remove exec call that happens to early
while reviewing rancher/ui#23993 i stumbled on a case where opening the pod logs
would cause an undefined error to crash the local ui server. `exec` is called
correctly from the `watchReconnect` observer which needs to happen first as it
parses the container name. If we call exec before we have a container name we
get an undefined error