Change the messaging and default creationtype
The backend was setting the default to legacy so that the legacy
template would still work when users migrate to the latest
version of the driver. This was causing the default value to
be legacy when creating a new vsphere template. Forcing
the backend to change the default will likely lead to problems
with the migration so we're making the change on the front end.

Ultimately we use the existing set default logic if a vsphere template
is being created (as opposed to edited) instead of just when the
creation type doesn't exist or is invalid.

We also changed the text for the legacy value to be friendlier.

rancher/rancher#24060