Setup observed property on init rather than default to object @ boot
`templateChanged` which loads the content of what we need for a MC app was
updated to watch a new property, `editable: { selectedTemplateUrl } `. The new
property was declared on the component as a object. Observers watching a static
object and that have never been fetched do not recompute. If we set the property
on init to the object it would have been declared with then additional sets on
the watched property cause the observer to recompute.

rancher/rancher#24057