Refactor role permissions validator
the old permissions validating was somewhat dumb in its checking, only looking
for local names of roles. This updates the check to loop through all the roles
and fetch there rules upon which we check for the existance of certain rules and
targets.

This also includes fixes for some broken links which happened as apart of the
previous change to this area.

This also fixes a broken translation

rancher/rancher#24341