Add support for using input-yaml in modals
When adding input-yaml to a modal it required the user to click on
the input before the value would be loaded and the editor resized.
I found a github issue which suggested making use of the isVisible
property to force a refresh so I added the use of it to force refreshing
to happen.

I also made it so the header didn't add a margin if the header wasn't
present.