Improve Vsphere UX around cloudinit/cloudconfig
There was confusion and misuse around cloudinit and cloudconfig.
This change switches things up so cloudinit is only used for the
legacy creation type and cloudconfig is used for all other creation
types.

Changes the textarea that was used for cloudconfig to
input-yaml to make it clear to users that they can enter yaml.

Switches the vmware cloud credential vcenter input to an input-url
so that we automatically strip the scheme from the input.

rancher/rancher#24314