add isCreatorMember for inital cluster or project create
we use this to hide the remove action button on the default row when creating
one fo the two mentioned resources. This is because the backend automatcially
sets up this binding on create. We display this fake member so the user doesnt
feel like they need to add themsevles as a member.