Better handle rejected promises in StatefulPromise
When the promise was rejected it took the error value and
set that as the value of the promise.

This lead to code expecting one shape to completely fail. This
now assigns the errorValue to an errorValue property on the
promise and the defaultValue to value when a promise is rejected.

rancher/rancher#24620