Refactor how cis scans are disabled.
This refactoring removed code duplication and
provides a more cohesive experience across all
of the pages that you can run a scan.

This will also now check and ensure the following
- there aren't running ClusterScans
- this isn't a windows cluster
- systemProject is available
- the cluster is active
- the action link is available

When the above isn't the case Run CIS Scan will not be present in
the cluster action drop downs and the Run CIS Scan but
 will be disabled on the CIS Scans and CIS Scan Detail
pages.

rancher/rancher#24759
rancher/rancher#25298