Remove empty userId field
Now, userId will only be set at a value if the member being added
is a user. Otherwise it will be omitted. Prior, userId was set
as an empty string. This interfered with validating that only a
group or user can be targetted. PRTBs can now target groups so
targeting a user should not be assumed.