Remove CIS skip button and total number
We're removing the CIS skip button because we decided the current UX
is poort and we don't want to spend a tremendous amount of effort
making a marginally better UX in the short time we have left.

We also removed the 19 of 59 from the main CIS page as the total was
misleading and wasn't actionable.

rancher/rancher#25863