Remove minimum k8s version range check for Cloud Provider clusters
Adds a new service which parses version from the various cloud provider version
list. I moved this to a new service rather than use form-versions because
form-versions is already fairly complicated with how it has to deal with RKE
Templates and unknown patch versions. It was simpler, cleaner, and faster to
move the CP cluster version parsing to a service and use new select because the
versions coming down do not include unknown patch versions. Addtionally going
this route allows us to not have to test all clusters for regressions, only CP ones.

rancher/rancher#26255