Make the default scheduledClusterScan.enabled to be a boolean
Enabled was being set to 'false' instead of false but this was being
masked by the execution of initScheduledClusterScan. When the
KDM values werent present initScheduledClusterScan was doing
an early exit and no longer masking the poor 'false' default value.

This will further improve the earlier solution for rancher/rancher#26996.