Adding a selected cis profile to the profile options if it's not present
When upgrading rancher the old 1.4 cis profiles were removed which prevented
the currently selected profile for scheduled scans to be selected.

This adds a selected profile to the options if it's not present allowing the
profile to be selected by default. If a different profile is selected and saved
the option will no longer be present. This option will also not be present in
the manual run scan modal.

rancher/rancher#27867
rancher/rancher#27866
rancher/rancher#27374