remove conflicting close action on page-header-project
when i ugpraded ember basic dropdown the mechanism we previously used to close the dropdown on click was removed and as a result I added the recommended method closing the dropdown.
the problem was that page-header-project and hover-drowndown worktogether to close the dropdown as well and the conflict was causing a full page refresh due to the close method.
ive dropped the conflicting action and instead moved the close logic to hover dropdown in the routeWillChange method since its knows its context in regards to the dropdown.
at this point the `a` in page header project will have been magically clicked and the component is done with publicDropDownApi so it can be set null.