hide local cluster from header & global clusters list if hide local true
if the new setting hide-local-cluster exists and is true we will hide the local cluster in the UI. this doesnt prevent api access.
additionally if the setting is true and a user attempts to navigate directly to a cluster or project route we will redirect them to the authenticated index route which redirects to global admin.

rancher/rancher#29325