Expose cluster reg token 4 cluster w/o nodes
For imported or custom clusters that never had a host register there was
never a way to get back to the registration command. This exposes a new
modal and button ont he cluster dashboard the allows the user to fetch
this command if the cluster doesn't have any nodes.

rancher/rancher#28548

Add import-command component the import command on imported clusters

this is the first step to exposing the command as an action in the modal for show command

Create CustomCommand component for modal show command

add custom input to show command modal

translations