Disabling all reference types on rke create for agentenvvar and marking variable name field as required
Since the reference types can't exist until the cluster is created I added an option to disable all reference types and use it during RKE create

Even though the backend doesn't indicate the variable name is required we're marking it as such because it casues cluster to fail. No validation will be done at this time.

rancher/rancher#31528