Refactor logic for showing CustomCommand and ImportCommand 4 env var
Since the env vars have to be saved on a cluster there is no easy way of dynamically updating the cluster command without trying to split the string and do all kinds of nasty things.
I refactored the logic for when we show the command for both Custom and Import clusters until after the cluster has been saved. This ensures the user is required to save the cluster before fetching the command on an edit action

rancher/rancher#31529