Added extra logic for excluding docker cluster from edit/upgrade
k3s clusters that do not have a config should not allow editing of items on the config.
A null k3s config indicates a docker installed cluster so I've added extra logic
to check if the config exists and if it does not we don't show the manage import cluster info component.

I also added logic to the final save method in cru cluster to only show the registration step if the cluster is pending
and not a null k3s config.

rancher/rancher#30977