catch labels case for gke config syncing
added labels to the diffUpstreamSpec comparision exclusion. we already check for equality (eg no changes) in the JSON string compare higher in the function so that handeld the labels that were equal but we need to exclude the labels from further comparison becuase if they are different in the user config, meaning the user changed them, then we know we need to send the entire labels object to update. If we didn't do this the comparison fails because we're only comparing the differences one way not both and removals on the map shorts the comparison functions further on in the function.

rancher/rancher#32433