Removing monitoring
Logging had already been removed from the navigation menu but the translation was still present. I removed the translation for completeness.

I left the routes and pages just in case we need to allow the customer to visit those pages even if it's unlikely. We will eventually fully remove ember so I don't think there's a reason to tidy up too much in this scenario.

https://github.com/rancher/dashboard/issues/10617